voice-agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg[binary]>=3.1
aiosqlite==0.20.0
pydantic-settings==2.7.0
cryptography==44.0.0
pgvector==0.3.6
openai==1.59.6
anthropic==0.42.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.157.0
httpx==0.28.1
python-multipart==0.0.20
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0
pypdf==5.1.0
python-docx==1.1.2
tiktoken==0.8.0
jinja2==3.1.4
